create components/PickUpLocationPage.js
